% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_tmod_mset}
\alias{add_tmod_mset}
\title{add_tmod_mset}
\usage{
add_tmod_mset(rnaseq, ...)
}
\arguments{
\item{rnaseq}{rna_analysis object}

\item{...}{name = value pairs of mset objects}
}
\value{
rnaseq_analysis object.
}
\description{
add_tmod_mset
}
