% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fCounts_res.R
\name{aggregate_fCounts_res}
\alias{aggregate_fCounts_res}
\title{Aggregate count files produced by featureCounts}
\usage{
aggregate_fCounts_res(count_files = NULL, sample_names = NULL)
}
\arguments{
\item{count_files}{vector of count file names}

\item{sample_names}{optional sample names.}
}
\description{
Aggregate count files produced by featureCounts
}
