% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_salmon_quant.R
\name{aggregate_salmon_quant}
\alias{aggregate_salmon_quant}
\title{Aggregate quant files produced by Salmon}
\usage{
aggregate_salmon_quant(quant_files = NULL, sample_names = NULL)
}
\arguments{
\item{quant_files}{vector of count file names}

\item{sample_names}{optional sample names.}
}
\description{
Aggregate quant files produced by Salmon
}
