% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{calculate_cooks_distance}
\alias{calculate_cooks_distance}
\title{Calculate Cook's Distance}
\usage{
calculate_cooks_distance(values)
}
\arguments{
\item{values}{A numeric vector of values for which to calculate Cook's Distance.}
}
\value{
A numeric vector of Cook's Distance values.
}
\description{
Calculate Cook's Distance
}
\examples{
values <- rnorm(100)
calculate_cooks_distance(values)
}
