% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{calculate_lof}
\alias{calculate_lof}
\title{Calculate Local Outlier Factor (LOF)}
\usage{
calculate_lof(values, k = 5)
}
\arguments{
\item{values}{A numeric vector of values for which to calculate LOF.}

\item{k}{The number of neighbors to use for calculating LOF. Defaults to 5.}
}
\value{
A numeric vector of LOF values.
}
\description{
Calculate Local Outlier Factor (LOF)
}
\examples{
values <- rnorm(100)
calculate_lof(values)
}
