% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{calculate_measures}
\alias{calculate_measures}
\title{Calculate Measures for Each Feature}
\usage{
calculate_measures(object)
}
\arguments{
\item{object}{An OmicsData object.}
}
\value{
The OmicsData object with calculated measures for each feature.
}
\description{
Calculate Measures for Each Feature
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
}
