% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qtrap.R
\name{combine_flagged_data}
\alias{combine_flagged_data}
\title{Combine Flagged Area and Height Data}
\usage{
combine_flagged_data(flagged_area, flagged_height, sample_id_col = "sample_id")
}
\arguments{
\item{flagged_area}{A tibble containing flagged area data.}

\item{flagged_height}{A tibble containing flagged height data.}

\item{sample_id_col}{Name of the column containing sample ID information.}
}
\value{
A combined tibble with the same dimensions and column names as the input data,
        containing "Too high in Blank" if flagged in either tibble.
}
\description{
This function combines flagged area and height data tibbles. If any feature is marked as "Too high in Blank"
in either tibble for a given sample, it will be marked as "Too high in Blank" in the combined tibble.
}
\examples{
\dontrun{
combined_data <- combine_flagged_data(flagged_area_data, flagged_height_data, "sample_id")
print(combined_data)
}
}
\author{
Yaoxiang Li
}
