% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{convert_to_binary_matrix}
\alias{convert_to_binary_matrix}
\title{Convert to Binary Matrix for UpSetR}
\usage{
convert_to_binary_matrix(upset_data)
}
\arguments{
\item{upset_data}{The data frame prepared for UpSet plot.}
}
\value{
A binary matrix for UpSetR.
}
\description{
Convert to Binary Matrix for UpSetR
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
upset_data <- prepare_upset_data(omics_data)
upset_matrix <- convert_to_binary_matrix(upset_data)
}
