% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_assigned_summary.R
\name{count_assigned_summary}
\alias{count_assigned_summary}
\title{Create a table integrating the summaries of STAR and featureCounts}
\usage{
count_assigned_summary(samples_dir = NULL)
}
\arguments{
\item{samples_dir}{directory with all samples; every sample is in a directory that contains "Log.final.out" and "counts.txt.summary"}
}
\description{
Create a table integrating the summaries of STAR and featureCounts
}
