% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{createOmicsData}
\alias{createOmicsData}
\title{Constructor for OmicsData}
\usage{
createOmicsData(data = NULL, n_samples = 100, n_features = 2000)
}
\arguments{
\item{data}{The data frame containing the data. If not provided, synthetic data will be generated.}

\item{n_samples}{The number of samples to generate if data is not provided. Defaults to 100.}

\item{n_features}{The number of features to generate if data is not provided. Defaults to 2000.}
}
\value{
An OmicsData object.
}
\description{
Constructor for OmicsData
}
\examples{
omics_data <- createOmicsData()
user_data <- generate_data_with_anomalies()
omics_data <- createOmicsData(data = user_data)
}
