% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Differential_expression for count data
Differential expression analysis based on limma}
\usage{
differential_expression(
  dge = NULL,
  design = NULL,
  contr_mat = NULL,
  out_dir = NULL,
  ...
)
}
\arguments{
\item{dge}{DGEList}

\item{design}{experimental design matrix}

\item{contr_mat}{contrast matrix; limma::makeContrasts()}

\item{out_dir}{output directory}

\item{...}{further arguments to limma::voomLmFit()}
}
\value{
A list with:
\enumerate{
\item fit; fit object
\item tt: list of differential expression tables
}
}
\description{
Differential_expression for count data
Differential expression analysis based on limma
}
