% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_DESeq.R
\name{differential_expression_DESeq}
\alias{differential_expression_DESeq}
\title{Differential expression analysis based on limma}
\usage{
differential_expression_DESeq(
  counts = NULL,
  sample_ann = NULL,
  design = NULL,
  contrast = NULL,
  type = "apeglm"
)
}
\arguments{
\item{counts}{genes-by-samples matrix of counts}

\item{sample_ann}{data.frame with sample annotation; row.names musst be identical to the colnames of counts}

\item{design}{design matrix}

\item{type}{algorithm for srhinkage see the documentation of lfcShrink()}

\item{contrasts}{contrasts}
}
\value{
A list with:
\enumerate{
\item fit; fit object
\item tt: list of differential expression tables
}
}
\description{
Differential expression analysis based on limma
}
