% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg2uniprot_10090.R
\docType{data}
\name{eg2uniprot_10090}
\alias{eg2uniprot_10090}
\title{eg2uniprot_10090
Maps between Entrez gene and Uniprot, , dowloaded from NCBI and UNIPROT on September 19, 2023.}
\format{
An object of class \code{data.frame} with 32658 rows and 2 columns.
}
\usage{
eg2uniprot_10090
}
\description{
eg2uniprot_10090
Maps between Entrez gene and Uniprot, , dowloaded from NCBI and UNIPROT on September 19, 2023.
}
\keyword{datasets}
