% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg2uniprot_9606.R
\docType{data}
\name{eg2uniprot_9606}
\alias{eg2uniprot_9606}
\title{eg2uniprot_9606
Maps between Entrez gene and Uniprot, , dowloaded from NCBI and UNIPROT on September 19, 2023.}
\format{
An object of class \code{data.frame} with 33485 rows and 2 columns.
}
\usage{
eg2uniprot_9606
}
\description{
eg2uniprot_9606
Maps between Entrez gene and Uniprot, , dowloaded from NCBI and UNIPROT on September 19, 2023.
}
\keyword{datasets}
