% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{ensure_enough_sets_for_upset}
\alias{ensure_enough_sets_for_upset}
\title{Ensure There Are Enough Sets for UpSet Plot}
\usage{
ensure_enough_sets_for_upset(upset_matrix)
}
\arguments{
\item{upset_matrix}{The binary matrix for UpSetR.}
}
\description{
Ensure There Are Enough Sets for UpSet Plot
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
upset_data <- prepare_upset_data(omics_data)
upset_matrix <- convert_to_binary_matrix(upset_data)
ensure_enough_sets_for_upset(upset_matrix)
}
