% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_normalize.R
\name{filter_and_normalize}
\alias{filter_and_normalize}
\title{Filter and normalize count data}
\usage{
filter_and_normalize(
  genes_by_samples_matrix = NULL,
  annotation = NULL,
  min_sample_perc = 0.2,
  min_cpm = 3,
  out_dir = "./",
  class = NULL,
  pal = NULL,
  top_pca = 500
)
}
\description{
Filter and normalize count data
}
