% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter and normalize count data}
\usage{
filter_counts(
  X = NULL,
  filter.by.cpm = TRUE,
  min.cpm = 3,
  min.counts = 10,
  subset.remove = NULL,
  min.samples = NULL
)
}
\arguments{
\item{X}{genes-by-samples count matrix}

\item{filter.by.cpm}{if TRUE filter by cmp rather then raw counts}

\item{min.cpm}{mininum count per million a gene must have in at least min_sample_perc percentage of samples}

\item{min.counts}{minimum number of counts, used only if filter.by.cpm is FALSE}

\item{subset.remove}{these samples will be removed}

\item{min.samples}{minimum number of samples}
}
\value{
filtered count matrix
}
\description{
Filter and normalize count data
}
