% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{flag_anomalies}
\alias{flag_anomalies}
\title{Flag Anomalies}
\usage{
flag_anomalies(object)
}
\arguments{
\item{object}{An OmicsData object.}
}
\value{
The OmicsData object with flagged anomalies.
}
\description{
Flag Anomalies
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
}
