% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_mapping_ncbi_10090_2023_09_19.R
\docType{data}
\name{gene_id_mapping_ncbi_10090_2023_09_19}
\alias{gene_id_mapping_ncbi_10090_2023_09_19}
\title{gene_id_mapping_ncbi_10090_2023_09_19
Maps among main identifiers, dowloaded from NCBI on September 19, 2023.}
\format{
An object of class \code{data.frame} with 289208 rows and 10 columns.
}
\usage{
gene_id_mapping_ncbi_10090_2023_09_19
}
\description{
gene_id_mapping_ncbi_10090_2023_09_19
Maps among main identifiers, dowloaded from NCBI on September 19, 2023.
}
\keyword{datasets}
