% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_mapping_ncbi_9606_2026_03_09.R
\docType{data}
\name{gene_id_mapping_ncbi_9606_2026_03_09}
\alias{gene_id_mapping_ncbi_9606_2026_03_09}
\title{gene_id_mapping_ncbi_9606_2026_03_09
Maps among main identifiers, dowloaded from NCBI on March 03, 2026.}
\format{
An object of class \code{data.frame} with 492994 rows and 10 columns.
}
\usage{
gene_id_mapping_ncbi_9606_2026_03_09
}
\description{
gene_id_mapping_ncbi_9606_2026_03_09
Maps among main identifiers, dowloaded from NCBI on March 03, 2026.
}
\keyword{datasets}
