% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_identifiers_2023_02_24.R
\docType{data}
\name{gene_identifiers_2023_02_24}
\alias{gene_identifiers_2023_02_24}
\title{gene_identifiers_2023_02_24
Maps among main identifiers, dowloaded from NCBI on Februrary 24, 2023.}
\format{
An object of class \code{data.frame} with 738402 rows and 11 columns.
}
\usage{
gene_identifiers_2023_02_24
}
\description{
gene_identifiers_2023_02_24
Maps among main identifiers, dowloaded from NCBI on Februrary 24, 2023.
}
\keyword{datasets}
