% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{generate_data_with_anomalies}
\alias{generate_data_with_anomalies}
\title{Generate High-Dimensional Data with Anomalies}
\usage{
generate_data_with_anomalies(n_samples = 100, n_features = 2000)
}
\arguments{
\item{n_samples}{The number of samples to generate. Defaults to 100.}

\item{n_features}{The number of features to generate. Defaults to 2000.}
}
\value{
A data frame containing the generated data with anomalies.
}
\description{
Generate High-Dimensional Data with Anomalies
}
\examples{
data <- generate_data_with_anomalies()
}
