% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_granges_heatmap.R
\name{heatmap_granges}
\alias{heatmap_granges}
\title{R function to create a genom-wide heatmap for one meta column for a list of GRanges}
\usage{
heatmap_granges(
  granges,
  meta_field,
  row_groups = NULL,
  keep_sites_present_in = floor(0.8 * length(granges)),
  color_breaks = c(0, 2, 6),
  colors = c("dodgerblue3", "white", "firebrick3"),
  window_width = 1e+06,
  upper_limit = 6,
  cluster_rows = FALSE,
  show_row_names = TRUE,
  show_column_names = FALSE,
  use_raster = TRUE,
  cluster_columns = FALSE,
  border = TRUE,
  legend_params = NULL,
  ...
)
}
\arguments{
\item{granges}{.}

\item{meta_field}{.}

\item{row_groups}{.}

\item{keep_sites_present_in}{.}

\item{color_breaks}{.}

\item{colors}{.}

\item{window_width}{.}

\item{upper_limit}{.}

\item{cluster_rows}{.}

\item{show_row_names}{.}

\item{show_column_names}{.}

\item{use_raster}{.}

\item{cluster_columns}{.}

\item{border}{.}

\item{legend_params}{.}

\item{...}{}
}
\value{
ComplexHeatmap heatmap
}
\description{
R function to create a genom-wide heatmap for one meta column for a list of GRanges
}
\examples{
.
}
