% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_map_2023_09_20.R
\docType{data}
\name{hgnc_map_2023_09_20}
\alias{hgnc_map_2023_09_20}
\title{hgnc_map_2023_09_20
Maps among main identifiers, dowloaded from HGNC on September 20, 2023.}
\format{
An object of class \code{data.frame} with 45526 rows and 6 columns.
}
\usage{
hgnc_map_2023_09_20
}
\description{
hgnc_map_2023_09_20
Maps among main identifiers, dowloaded from HGNC on September 20, 2023.
}
\keyword{datasets}
