% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_analysis_class.R
\name{init_rnaseq_analysis}
\alias{init_rnaseq_analysis}
\title{Initialise RNAseq analysis}
\usage{
init_rnaseq_analysis(deseq2)
}
\arguments{
\item{deseq2}{Named list of deseq2 results}
}
\value{
rnaseq_analysis object.
}
\description{
Initialise RNAseq analysis
}
