% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{initialize_results_df}
\alias{initialize_results_df}
\title{Initialize Results Data Frame}
\usage{
initialize_results_df(data)
}
\arguments{
\item{data}{The data frame containing the generated data.}
}
\value{
A data frame initialized with columns for anomaly measures.
}
\description{
Initialize Results Data Frame
}
\examples{
data <- generate_data_with_anomalies()
anomaly_measures <- initialize_results_df(data)
}
