% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{internal_standard_normalize}
\alias{internal_standard_normalize}
\title{Internal Standard Normalize}
\usage{
internal_standard_normalize(area_data, istd_area_data)
}
\arguments{
\item{area_data}{A data frame containing the area data.}

\item{istd_area_data}{A data frame containing the internal standard area data.}
}
\value{
A data frame with normalized data.
}
\description{
This function performs internal standard normalization.
}
\examples{
area_data <- data.frame(sample_id = c("S1", "S2", "S3"), A = 1:3, B = 4:6)
istd_area_data <- data.frame(sample_id = c("S1", "S2", "S3"), A = 1:3, B = 4:6)
normalized_data <- internal_standard_normalize(area_data, istd_area_data)
}
