% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_is_normalize.R
\name{is_normalize}
\alias{is_normalize}
\title{Normalize loaded tibble with IS tibble}
\usage{
is_normalize(loaded_tibble, is_tibble)
}
\arguments{
\item{loaded_tibble}{A tibble loaded by 'load_peak_area' or similar function.}

\item{is_tibble}{A tibble loaded by 'load_is_area' or similar function.}
}
\value{
A tibble that is normalized by the IS tibble.
}
\description{
Normalize loaded tibble with IS tibble
}
\author{
Yaoxiang Li \email{yl814@georgetown.edu}

Georgetown University, USA

License: GPL (>= 3)
}
