% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_from_table.R
\name{jaccard_from_table}
\alias{jaccard_from_table}
\title{Jaccard index from a table of intersections}
\usage{
jaccard_from_table(x = NULL)
}
\arguments{
\item{x}{table of intersections between values of two categorical variables (intersections)}
}
\description{
Jaccard index from a table of intersections
}
