% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_load_is_area.R
\name{load_is_area}
\alias{load_is_area}
\title{Load and clean IS area data}
\usage{
load_is_area(is_area_path, delim = "\\t", na = c("N/A", "Unknown"))
}
\arguments{
\item{is_area_path}{Character string specifying the path to the IS area data file.}

\item{delim}{Character string indicating the delimiter used in the file.}

\item{na}{Character vector of strings to be treated as NA. Defaults to c("N/A", "Unknown").}
}
\value{
A tibble containing the cleaned IS area data.
}
\description{
Load and clean IS area data
}
\author{
Yaoxiang Li \email{yl814@georgetown.edu}

Georgetown University, USA

License: GPL (>= 3)
}
