% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{load_peak_area}
\alias{load_peak_area}
\title{Load and clean peak area data}
\usage{
load_peak_area(peak_area_path, delim = "\\t", na = c("N/A", "Unknown"))
}
\arguments{
\item{peak_area_path}{Character string specifying the path to the peak area data file.}

\item{delim}{Character string indicating the delimiter used in the file. Defaults to "\t".}

\item{na}{Character vector of strings to be treated as NA. Defaults to c("N/A", "Unknown").}
}
\value{
A tibble containing the cleaned peak area data.
}
\description{
Load and clean peak area data
}
