% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{ms1_annotation}
\alias{ms1_annotation}
\title{MS1 Annotation}
\usage{
ms1_annotation(
  AnnotaData,
  masstole = 0.05,
  toleUnit = 1,
  annotaDB = "metlin",
  ionMode = "pos",
  adducts = NULL
)
}
\description{
MS1 Annotation
}
