% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{nist_rsd}
\alias{nist_rsd}
\title{Calculate the relative standard deviation for NIST samples}
\usage{
nist_rsd(normalized_tibble, nist_name = "NIST")
}
\arguments{
\item{normalized_tibble}{A tibble normalized by 'is_normalize' or similar function.}

\item{nist_name}{Character string specifying the name of the NIST sample. Defaults to "NIST".}
}
\value{
A tibble containing the mean, standard deviation, and coefficient of variation for NIST samples.
}
\description{
Calculate the relative standard deviation for NIST samples
}
