% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Filter and normalize count data}
\usage{
normalize_counts(
  X = NULL,
  gene.annotation = NULL,
  sample.annotation = NULL,
  design = NULL
)
}
\arguments{
\item{X}{genes-by-samples count matrix}

\item{gene.annotation}{data.frame with gene annotations}

\item{sample.annotation}{data.frame with sample annotations}

\item{design}{experimental design matrix}
}
\value{
a list with the following elements
\enumerate{
\item dge DGEList
\item lcpm log counts per million matrix
\item dds DESeq2 object
\item vsd DESeq2 transform object, resulting from vst(, blind=TRUE) for exploratory analysis
\item vsdbf DESeq2 transform object, resulting from vst(..., blind=FALSE) for downstream analysis
}
}
\description{
Filter and normalize count data
}
