% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\docType{class}
\name{OmicsData-class}
\alias{OmicsData-class}
\title{OmicsData Class}
\description{
A class to represent and analyze high-dimensional omics data.
}
\section{Slots}{

\describe{
\item{\code{data}}{The data frame containing the generated data with anomalies.}

\item{\code{anomaly_measures}}{The data frame containing the calculated anomaly measures.}

\item{\code{thresholds}}{The list of thresholds for anomaly detection.}
}}

