% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_batch_effect.R
\name{perform_pvca_batch_assessment}
\alias{perform_pvca_batch_assessment}
\title{Perform Principal Variance Component Analysis (PVCA) for Batch Effect Assessment}
\usage{
perform_pvca_batch_assessment(
  data_matrix,
  sample_info,
  batch_effects,
  threshold
)
}
\arguments{
\item{data_matrix}{A data frame or matrix where rows represent features and columns represent samples.}

\item{sample_info}{A data frame containing sample information with rows matching the columns of `data_matrix`.}

\item{batch_effects}{A character vector of column names in `sample_info` that represent batch effects.}

\item{threshold}{A numeric value between 0 and 1 to specify the PVCA threshold.}
}
\value{
A PVCA object containing the results of the batch effect assessment.
}
\description{
This function performs Principal Variance Component Analysis (PVCA) to assess batch effects in the dataset.
}
\examples{
pvca_results <- perform_pvca_batch_assessment(data_matrix, sample_info, c("batch", "dose", "time"), 0.6)
}
