% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{plot_distribution_measures}
\alias{plot_distribution_measures}
\title{Plot Distribution Measures}
\usage{
plot_distribution_measures(object)
}
\arguments{
\item{object}{An OmicsData object.}
}
\description{
Plot Distribution Measures
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
plot_distribution_measures(omics_data)
}
