% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lib_size.R
\name{plot_lib_size}
\alias{plot_lib_size}
\title{Plot library size}
\usage{
plot_lib_size(X = NULL, col = NULL, sort = FALSE)
}
\arguments{
\item{X}{genes-by-samples count matrix}

\item{col}{vector of colors; this must match the number of columns of X}

\item{sort}{whether to sort the samples by library size}
}
\value{
ggplot2 object
}
\description{
Plot library size
}
