% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RLE.R
\name{plot_RLE}
\alias{plot_RLE}
\title{RLE plot}
\usage{
plot_RLE(X = NULL, col = NULL)
}
\arguments{
\item{X}{Relative log expression}

\item{col}{vector of colors; this must match the number of columns of X}
}
\value{
ggplot2 object
}
\description{
RLE plot
}
