% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{plot_sample_measures}
\alias{plot_sample_measures}
\title{Plot Sample Measures}
\usage{
plot_sample_measures(object)
}
\arguments{
\item{object}{An OmicsData object.}
}
\description{
Plot Sample Measures
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
plot_sample_measures(omics_data)
}
