% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_tmod_panelplot}
\alias{plot_tmod_panelplot}
\title{Plots tmod pamelPlot}
\usage{
plot_tmod_panelplot(
  rnaseq,
  mset_name = NULL,
  tests = rnaseq$tests,
  pie.style = "pie",
  filter.rows.pval = 0.001,
  ...
)
}
\arguments{
\item{rnaseq}{rnaseq_analysis object}
}
\value{
rnaseq_analysis object
}
\description{
Plots tmod pamelPlot
}
