% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Volcano plot}
\usage{
plot_volcano(
  df,
  p_adj_col = "padj",
  lfc_col = "log2FoldChange",
  label_col = "mgi_symbol",
  p_adj_treshold = 0.05,
  lfc_treshold = 1
)
}
\description{
Plot Volcano plot
}
