% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_feature.R
\name{prepare_upset_data}
\alias{prepare_upset_data}
\title{Prepare Data for UpSet Plot}
\usage{
prepare_upset_data(object)
}
\arguments{
\item{object}{An OmicsData object.}
}
\value{
A data frame prepared for UpSet plot.
}
\description{
Prepare Data for UpSet Plot
}
\examples{
omics_data <- createOmicsData()
omics_data <- calculate_measures(omics_data)
omics_data <- flag_anomalies(omics_data)
upset_data <- prepare_upset_data(omics_data)
}
