% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qc_rsd.R
\name{qc_rsd}
\alias{qc_rsd}
\title{Calculate the relative standard deviation for QC samples}
\usage{
qc_rsd(normalized_tibble, qc_name = "Pooled QC")
}
\arguments{
\item{normalized_tibble}{A tibble normalized by 'is_normalize' or similar function.}

\item{qc_name}{Character string specifying the name of the QC sample. Defaults to "Pooled QC".}
}
\value{
A tibble containing the mean, standard deviation, and coefficient of variation for QC samples.
}
\description{
Calculate the relative standard deviation for QC samples
}
\author{
Yaoxiang Li \email{yl814@georgetown.edu}

Georgetown University, USA

License: GPL (>= 3)
}
