% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_star_final_log.R
\name{read_star_final_log}
\alias{read_star_final_log}
\title{Read the output of STAR}
\usage{
read_star_final_log(file = NULL)
}
\arguments{
\item{file}{"Log.final.out"}
}
\description{
Read the output of STAR
}
