% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run_tmod_cerno}
\alias{run_tmod_cerno}
\title{Runs tmod CERNO test for each tmod mset}
\usage{
run_tmod_cerno(rnaseq, ...)
}
\arguments{
\item{rnaseq}{rnaseq_analysis object}
}
\value{
rnaseq_analysis object
}
\description{
Runs tmod CERNO test for each tmod mset
}
