% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{save_evidence_plots}
\alias{save_evidence_plots}
\title{Plots tmod evidence plots}
\usage{
save_evidence_plots(rnaseq, modules_to_detail, out_dir = "images/")
}
\arguments{
\item{rnaseq}{rnaseq_analysis object}

\item{modules_to_detail}{list(mset_name = vec_of_modules)}
}
\description{
Plots tmod evidence plots
}
