% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose DataFrame}
\usage{
transpose_df(df)
}
\arguments{
\item{df}{A data frame to be transposed.}
}
\value{
A transposed data frame.
}
\description{
This function transposes a data frame by converting rows to columns and columns to rows.
The first column is assumed to contain the variable names, and the remaining columns contain the values.
}
\examples{
# Example usage:
df <- data.frame(
  ID = c("A", "B", "C"),
  Var1 = c(1, 2, 3),
  Var2 = c(4, 5, 6)
)
transpose_df(df)
}
