% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{differential_expression}
\usage{
differential_expression(
  dge = NULL,
  design = NULL,
  contr_mat = NULL,
  out_dir = "./",
  top_genes = 10,
  top_genes_column = "symbol"
)
}
\description{
differential_expression
}
