% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{differential_expression}
\usage{
differential_expression(dge = NULL, design = NULL, contr_mat = NULL)
}
\description{
differential_expression
}
