% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Differential_expression for count data
Differential expression analysis based on limma}
\usage{
differential_expression(
  dge = NULL,
  design = NULL,
  contr_mat = NULL,
  out_dir = NULL,
  top_genes = 10,
  top_genes_column = "symbol",
  block = NULL
)
}
\arguments{
\item{dge}{DGEList}

\item{design}{experimental design matrix}

\item{contr_mat}{contrast matrix; see limma::makeContrasts()}

\item{out_dir}{output directory}

\item{top_genes}{number of top genes to label in the volcano plots}

\item{top_genes_column}{column with gene labels for volcano plots}
}
\value{
A list with:
\enumerate{
  \item fit; fit object
  \item tt: list of differential expression tables
}
}
\description{
Differential_expression for count data
Differential expression analysis based on limma
}
